gateway/requirements.txt
fastapi
uvicorn
nameko-grpc
python-dotenv 